backend/requirements.txt
Flask
Flask-Cors
python-dotenv
google-genai
pydantic
